Update PopUpList.cpp
The height of the listitems in the dropdown shouldn't be determined by the font-size. It should be taken from the list it encapsulates. See below code for the reason:

DropList dl;
dl.SetLineCy(32);
dl.Add(String("1")).Add(String("2"));
dl.SetLineCy(0, 48);

in this case the first row will be drawn with height 48px and second with (on my machine) height 15px. This change ensure the second line will be drawn with height 32px.