CtrlLib, Doc: Tutorial(Gui24) correct comment and copywriting strings (#243) for zh-cn and zh-tw locale
Correct comment and copywriting strings for zh-cn and zh-tw locale.

As a follow up of #241. The documentation related file is also updated.

Note: I didn't update the attached image resources in the doc (tpp) file, feel free to directly adding or modifying this PR to update these images if preferred.